old-tests/submission/Infraware/Forms/contents/Forms/option_form.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>option_form</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype">
		  <option value="1"> Miner </option>
		  <option value="2"> Puffer </option>
		  <option value="3"> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		var option = select.item(2);
		var form = option.form;

		if (typeof(form) == "object") {
			test(function() {
				assert_equals(form.name, 'input_form', "form attribute is not exist.");
			});
		} else {
			test(function() {
				assert_unreached("form attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
